chore(v8/deps): Deduplicate yarn.lock
