+++ .github/workflows/build.yml
@@ -738,8 +738,8 @@ jobs:
           node_version: ${{ matrix.node == 14 && '14' || '' }}
 
       - name: Overwrite typescript version
-        if: matrix.typescript
-        run: node ./scripts/use-ts-version.js ${{ matrix.typescript }}
+        if: matrix.typescript == '3.8'
+        run: node ./scripts/use-ts-3_8.js
         working-directory: dev-packages/node-integration-tests
 
       - name: Run integration tests
